# api/openapi/v1.yaml
openapi: "3.1.0"
info:
  title: MIDAS API
  version: "1.0.0"
  description: >
    Authority governance engine for autonomous decisions.
    Every evaluation produces exactly one outcome and one tamper-evident audit envelope.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Static bearer token. Configured via MIDAS_AUTH_TOKENS or midas.yaml auth.tokens.
        Not required when auth.mode=open (development only).
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SubmitDriftDefinitionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        submitted_by:
          type: string
          description: Optional body fallback when the authenticated principal is not present.
        reason:
          type: string
    ApproveDriftDefinitionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        approved_by:
          type: string
          description: Optional body fallback when the authenticated principal is not present.
    RejectDriftDefinitionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        rejected_by:
          type: string
        reason:
          type: string
    DeprecateDriftDefinitionRequest:
      type: object
      description: >
        DriftDefinition deprecation request. Successor information is
        optional; when supplied it is persisted on the row's
        successor_definition_id and successor_version fields. The
        operator-supplied reason is captured in the control-audit
        record only.
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        deprecated_by:
          type: string
        reason:
          type: string
        successor_definition_id:
          type: string
        successor_version:
          type: integer
          minimum: 0
    RetireDriftDefinitionRequest:
      type: object
      description: >
        DriftDefinition retirement request. Allowed source states:
        draft, review, active, deprecated. Retired is terminal.
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        retired_by:
          type: string
        reason:
          type: string
...
    DriftType:
      type: string
      description: >
        V1 drift taxonomy. Population, data, prediction, performance,
        and concept drift are deferred to V2 and MUST NOT appear here.
      enum:
        - invocation
        - outcome
        - confidence
        - latency
        - evidence
        - authority
        - policy
        - coverage
        - process_path
    DriftDefinitionStatus:
      type: string
      enum:
        - draft
        - review
        - active
        - deprecated
        - retired
...
    DriftBaselineStrategy:
      type: string
      description: >
        V1 baseline strategies. champion_challenger and
        seasonality_aware are deferred to V2 and MUST NOT appear here.
      enum:
        - rolling
        - fixed_governed
        - previous_equivalent
        - since_last_governed_change
        - manually_pinned
    DriftCadence:
      type: string
      enum:
        - minute
        - hour
        - day
        - week
    DriftOrigin:
      type: string
      enum:
        - manual
        - inferred
        - auto_instrumented
    DriftTargetEntityKind:
      type: string
      description: >
        The nine governed entity kinds a DriftDefinition may target.
      enum:
        - business_service
        - capability
        - process
        - decision_surface
        - ai_system
        - ai_system_binding
        - agent
        - authority_profile
        - authority_grant
    DriftThresholdDirection:
      type: string
      enum:
        - ascending
        - descending
    DriftTargetRef:
      type: object
      description: Discriminated reference to a governed entity.
      required: [kind, id]
      properties:
        kind:
          $ref: "#/components/schemas/DriftTargetEntityKind"
        id:
          type: string
    DriftMetricDefinition:
      type: object
      description: Embedded metric inside a DriftDefinition revision.
      required:
        - metric_id
        - drift_type
        - baseline_strategy
        - window_seconds
        - cadence
        - threshold_direction
        - warning_threshold
        - breached_threshold
      properties:
        metric_id:
          type: string
        drift_type:
          $ref: "#/components/schemas/DriftType"
        baseline_strategy:
          $ref: "#/components/schemas/DriftBaselineStrategy"
        baseline_window_seconds:
          type: integer
          minimum: 0
        window_seconds:
          type: integer
          minimum: 1
        cadence:
          $ref: "#/components/schemas/DriftCadence"
        warning_threshold:
          type: number
        breached_threshold:
          type: number
        threshold_direction:
          $ref: "#/components/schemas/DriftThresholdDirection"
        governance_expectation_ref:
          type: string
        governance_expectation_ver:
          type: integer
          minimum: 0
        description:
          type: string
    DriftDefinition:
      type: object
      description: >
        Governed, versioned drift specification. Composite (id, version)
        identity. Apply persists Status='review'; activation lives in a
        future approval-endpoint tranche.
      required:
        - id
        - version
        - name
        - status
        - effective_date
        - business_owner
        - technical_owner
        - target
        - metrics
        - origin
        - managed
        - created_at
        - updated_at
      properties:
        id:
          type: string
        version:
          type: integer
          minimum: 1
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/DriftDefinitionStatus"
        effective_date:
          type: string
          format: date-time
        effective_until:
          type: string
          format: date-time
          nullable: true
        retired_at:
          type: string
          format: date-time
          nullable: true
        business_owner:
          type: string
        technical_owner:
          type: string
        target:
          $ref: "#/components/schemas/DriftTargetRef"
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/DriftMetricDefinition"
        origin:
          $ref: "#/components/schemas/DriftOrigin"
        managed:
          type: boolean
        replaces:
          type: string
        successor_definition_id:
          type: string
        successor_version:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
        approved_by:
          type: string
        approved_at:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /v1/controlplane/drift_definitions/{id}/submit:
    post:
      operationId: submitDriftDefinition
      summary: Submit a DriftDefinition revision for review
      description: >
        Transitions a specific DriftDefinition revision from draft to
        review. Requires platform.operator or platform.admin role
        (permission `drift_definition:submit`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDriftDefinitionRequest"
      responses:
        "200":
          description: DriftDefinition transitioned to review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriftDefinition"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission (requires drift_definition:submit)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: DriftDefinition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: DriftDefinition is not in draft state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Approval service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/controlplane/drift_definitions/{id}/approve:
    post:
      operationId: approveDriftDefinition
      summary: Approve a DriftDefinition revision in review state
      description: >
        Transitions a specific DriftDefinition revision from review to
        active. Maker-checker is enforced at the service layer: the
        approver must not equal the revision's `created_by` field
        (case-insensitive trim). Requires governance.approver or
        platform.admin role (permission `drift_definition:approve`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDriftDefinitionRequest"
      responses:
        "200":
          description: DriftDefinition approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriftDefinition"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission or maker-checker violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: DriftDefinition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: DriftDefinition is not in review state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Approval service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/controlplane/drift_definitions/{id}/reject:
    post:
      operationId: rejectDriftDefinition
      summary: Reject a DriftDefinition revision back to draft
      description: >
        Transitions a specific DriftDefinition revision from review back
        to draft. Approved_by and approved_at are not modified
        (review-state revisions have neither set). The operator-supplied
        reason is captured in the control-audit record only. Requires
        governance.approver or platform.admin role (permission
        `drift_definition:reject`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectDriftDefinitionRequest"
      responses:
        "200":
          description: DriftDefinition transitioned to draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriftDefinition"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission (requires drift_definition:reject)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: DriftDefinition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: DriftDefinition is not in review state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Approval service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/controlplane/drift_definitions/{id}/deprecate:
    post:
      operationId: deprecateDriftDefinition
      summary: Deprecate an active DriftDefinition revision
      description: >
        Transitions a specific DriftDefinition revision from active to
        deprecated. Successor information is optional; when supplied
        it is persisted on the row's `successor_definition_id` and
        `successor_version` fields. The operator-supplied reason is
        captured in the control-audit record only. Requires
        platform.admin role (permission `drift_definition:deprecate`).
        Cross-version atomic deprecation (auto-deprecate prior active
        when activating a new revision) is intentionally NOT performed
        — operators must call this endpoint explicitly.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeprecateDriftDefinitionRequest"
      responses:
        "200":
          description: DriftDefinition deprecated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriftDefinition"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission (requires drift_definition:deprecate)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: DriftDefinition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: DriftDefinition is not in active state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Approval service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/controlplane/drift_definitions/{id}/retire:
    post:
      operationId: retireDriftDefinition
      summary: Retire a DriftDefinition revision
      description: >
        Transitions a DriftDefinition revision to retired. Allowed
        source states: draft, review, active, deprecated. Retired is
        terminal — already-retired revisions are rejected with 409.
        Requires platform.admin role (permission
        `drift_definition:retire`).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetireDriftDefinitionRequest"
      responses:
        "200":
          description: DriftDefinition retired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriftDefinition"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission (requires drift_definition:retire)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: DriftDefinition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: DriftDefinition is already retired (terminal)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Approval service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...